Modified functionality to improve calculation of total cost.

+++ task3/task3a_b/src/main/java/com/task3a_b/Transport.java
@@ -27,7 +27,7 @@ public abstract class Transport implements Serializable{
         return transportationSpeed * distance;
     }
 
-    public double calculate_delivery_time(float distance){
+    public double calculate_delivery_time(double distance){
         return distance/transportationSpeed;
     }
